+++ database/migrations/2017_10_31_142226_create_transactions_table.php
@@ -22,7 +22,7 @@ class CreateTransactionsTable extends Migration
                 Debit real,
                 Credit real,
                 Type text NOT NULL,
-                Account_Number int NOT NULL,
+                Account_Number bigint NOT NULL,
                 Sender_Acc_No int,
                 Beneficiary_Acc_No int,
                 Foreign Key(Account_Number) References accounts (Account_Number)
